student360/fixbug_api_student_list

+++ student360/openerp-resource-server/src/main/java/openerp/openerpresourceserver/entity/ContestSubmission.java
@@ -31,7 +31,7 @@ public class ContestSubmission implements Serializable {
 
     private String userSubmissionId;
 
-    private int point;
+    private long point;
 
     private String status;
